requirements.txt
asgiref==3.4.1
backports.zoneinfo==0.2.1
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.9
colorama==0.4.4
coreapi==2.3.3
coreschema==0.0.4
cryptography==36.0.2
datetime-truncate==1.1.1
decorator==4.4.2
defusedxml==0.7.1
diff-match-patch==20200713
dj-database-url==0.5.0
Django==3.2.9
django-bower==5.2.0
django-debug-toolbar==3.2.2
django-environ==0.8.1
django-import-export==2.8.0
django-jazzmin==2.4.9
django-js-asset==2.0.0
django-model-utils==4.2.0
django-modeladmin-reorder==0.3.1
django-mptt==0.13.4
django-multiselectfield==0.1.12
django-nested-inline==0.4.4
django-nvd3==0.9.7
django-rest-passwordreset==1.2.1
djangorestframework==3.12.4
djcacheutils==3.0.0
drf-yasg==1.20.0
et-xmlfile==1.1.0
idna==3.3
imageio==2.14.0
imageio-ffmpeg==0.4.5
inflection==0.5.1
install==1.3.5
itypes==1.2.0
Jinja2==3.0.3
MarkupPy==1.14
MarkupSafe==2.0.1
moviepy==1.0.3
numpy==1.22.1
odfpy==1.4.1
openapi-codec==1.3.2
openpyxl==3.0.9
packaging==21.3
Pillow==9.0.0
proglog==0.1.9
psycopg2-binary==2.9.3
pycparser==2.21
pyparsing==3.0.6
python-dateutil==2.8.2
python-memcached==1.59
python-nvd3==0.14.2
python-slugify==1.1.4
pytz==2021.3
PyYAML==6.0
requests==2.26.0
ruamel.yaml==0.17.17
ruamel.yaml.clib==0.2.6
simplejson==3.17.6
six==1.16.0
sqlparse==0.4.2
tablib==3.2.1
tqdm==4.62.3
Unidecode==1.3.4
uritemplate==4.1.1
urllib3==1.26.7
whitenoise==5.3.0
xlrd==2.0.1
xlwt==1.3.0
